# Conflicts:
#	pom.xml
#	projet-ose-react/package-lock.json
#	projet-ose-react/package.json
#	projet-ose-react/src/App.test.tsx
#	projet-ose-react/src/App.tsx
#	projet-ose-react/src/api/InterOfferJobAPI.ts
#	projet-ose-react/src/components/common/ConnectForm.tsx
#	projet-ose-react/src/components/common/InternOfferList.tsx
#	projet-ose-react/src/components/common/InternshipOfferForm.tsx
#	projet-ose-react/src/components/common/InternshipOfferModal.tsx
#	projet-ose-react/src/components/common/inscriptionEmployeur.tsx
#	projet-ose-react/src/components/common/student/form/StudentInscriptionForm.tsx
#	projet-ose-react/src/components/common/testBackEndConnection.tsx
#	projet-ose-react/src/i18n.ts
#	projet-ose-react/src/index.tsx
#	projet-ose-react/src/pages/ErrorPage.tsx
#	projet-ose-react/src/pages/HomePage.tsx
#	projet-ose-react/src/pages/LoginPage.tsx
#	projet-ose-react/src/pages/management/offers/InternshipOfferPage.tsx
#	projet-ose-react/src/pages/student/InternshipOfferPage.tsx
#	projet-ose-react/src/pages/student/StudentInscriptionPage.tsx
#	projet-ose-react/src/tailwind.css
#	projet-ose-react/src/utils/switcher.js
#	projet-ose-react/tailwind.config.js
#	readme.md
#	src/main/java/com/sap/ose/projetose/ProjetOseApplication.java
#	src/main/java/com/sap/ose/projetose/controller/EtudiantController.java
#	src/main/java/com/sap/ose/projetose/controller/InternOfferController.java
#	src/main/java/com/sap/ose/projetose/controller/InternshipCandidatesController.java
#	src/main/java/com/sap/ose/projetose/dto/EmployeurDto.java
#	src/main/java/com/sap/ose/projetose/dto/InternOfferDto.java
#	src/main/java/com/sap/ose/projetose/dto/UtilisateurDto.java
#	src/main/java/com/sap/ose/projetose/modeles/Employeur.java
#	src/main/java/com/sap/ose/projetose/modeles/Etudiant.java
#	src/main/java/com/sap/ose/projetose/modeles/File.java
#	src/main/java/com/sap/ose/projetose/modeles/InternOffer.java
#	src/main/java/com/sap/ose/projetose/modeles/Internshipmanager.java
#	src/main/java/com/sap/ose/projetose/modeles/OfferReviewRequest.java
#	src/main/java/com/sap/ose/projetose/repository/EmployeurRepository.java
#	src/main/java/com/sap/ose/projetose/repository/EtudiantRepository.java
#	src/main/java/com/sap/ose/projetose/repository/InternOfferRepository.java
#	src/main/java/com/sap/ose/projetose/service/EmployeurService.java
#	src/main/java/com/sap/ose/projetose/service/EtudiantService.java
#	src/main/java/com/sap/ose/projetose/service/FileService.java
#	src/main/java/com/sap/ose/projetose/service/InternOfferService.java
#	src/main/java/com/sap/ose/projetose/service/InternshipmanagerService.java
#	src/main/java/com/sap/ose/projetose/service/OseService.java
#	src/test/java/com/sap/ose/projetose/controller/EtudiantControllerTest.java

+++ src/main/java/com/sap/ose/projetose/dto/UtilisateurDto.java
@@ -7,13 +7,13 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public abstract class UtilisateurDto {
-private long id;
+    private long id;
     private String nom;
     private String prenom;
     private String phone;
     private String email;
 
-    public UtilisateurDto(int id, String nom, String prenom, String phone, String email) {
+    public UtilisateurDto(long id, String nom, String prenom, String phone, String email) {
         this.nom = nom;
         this.prenom = prenom;
         this.phone = phone;
@@ -27,4 +27,6 @@ private long id;
         this.phone = phone;
         this.email = email;
     }
+
+
 }
